# Add `GetTokenRangesForPartition` method for partition ring